+++ .github/workflows/workshop-1pass-esc.yml
@@ -1,13 +1,8 @@
 name: workshop-1pass-esc
 
-
 on:
   pull_request:
     branches: [ "main" ]
-# on:
-#   push:
-#     paths:
-#       - 'infra/*'
 
 jobs:
   workshop-1pass-esc:
