+++ app/main.go
@@ -2,49 +2,37 @@ package main
 
 import (
 	"context"
-	"crypto/rand"
-	"encoding/base64"
 	"encoding/json"
 	"fmt"
-	"io"
-	"io/ioutil"
 	"log"
 	"net/http"
 	"os"
-	"time"
 
-	"regexp"
 	"strings"
-	"unicode"
 
 	"github.com/google/generative-ai-go/genai"
+	"golang.org/x/net/html"
 	"golang.org/x/oauth2"
 	"golang.org/x/oauth2/google"
 	"google.golang.org/api/option"
 )
 
-func New() http.Handler {
-	mux := http.NewServeMux()
-	// Roots
-	mux.Handle("/", http.FileServer(http.Dir("templates/")))
-
-	// OAuth with Google
-	mux.HandleFunc("/auth/google/login", oauthGoogleLogin)
-	mux.HandleFunc("/auth/google/callback", oauthGoogleCallback)
-
-	// App endpoint
-	mux.HandleFunc("/submit", submitEndpoint)
-
-	return mux
+type Profile struct {
+	ID            string `json:"id"`
+	Email         string `json:"email"`
+	VerifiedEmail bool   `json:"verified_email"`
+	Name          string `json:"name"`
+	GivenName     string `json:"given_name"`
+	FamilyName    string `json:"family_name"`
+	Picture       string `json:"picture"`
+	Locale        string `json:"locale"`
+	Hd            string `json:"hd"`
+	AccessToken   string `json:"accessToken"`
 }
 
-func getEnvOrDefault(envVarName, defaultValue string) string {
-	value := os.Getenv(envVarName)
-	if value == "" {
-		log.Printf("Environment variable %s not set, using default value %s", envVarName, defaultValue)
-		return defaultValue
-	}
-	return value
+type BuzzForm struct {
+	A string `json:"a"`
+	P string `json:"p"`
 }
 
 const (
@@ -53,41 +41,14 @@ const (
 	OAUTH_API = "https://www.googleapis.com/oauth2/v2/userinfo?access_token="
 )
 
-func parseResponse(resp *genai.GenerateContentResponse) (string, error) {
+func New() http.Handler {
+	mux := http.NewServeMux()
+	mux.Handle("/", http.FileServer(http.Dir("templates/")))
+	mux.HandleFunc("/auth/google/login", loginHandler)
+	mux.HandleFunc("/app", appHandler)
+	mux.HandleFunc("/submit", submitHandler)
 
-	var formattedContent string
-
-	for _, cand := range resp.Candidates {
-		if cand.Content != nil {
-			for _, part := range cand.Content.Parts {
-				formattedContent += fmt.Sprintf("%s\n", part)
-			}
-		} else {
-			return "", fmt.Errorf("no content found in response")
-		}
-	}
-	return formattedContent, nil
-}
-
-func cleanInput(input string) string {
-	// Define a regex pattern to match non-alphanumeric characters and diacritics
-	pattern := regexp.MustCompile(`[^a-zA-Z\\s]+`)
-
-	// Replace non-alphanumeric characters and diacritics with an empty string
-	cleaned := pattern.ReplaceAllStringFunc(input, func(s string) string {
-		var result []rune
-		for _, r := range s {
-			// Check if the rune is alphanumeric or whitespace
-			if unicode.IsLetter(r) || unicode.IsSpace(r) {
-				result = append(result, r)
-			}
-		}
-		return string(result)
-	})
-
-	// Remove extra whitespaces
-	cleaned = strings.Join(strings.Fields(cleaned), " ")
-	return cleaned
+	return mux
 }
 
 // main is the entry point for the application.
@@ -103,8 +64,6 @@ func main() {
 
 	if err := server.ListenAndServe(); err != http.ErrServerClosed {
 		log.Printf("%v", err)
-	} else {
-		log.Println("Server closed!")
 	}
 }
 
@@ -115,45 +74,7 @@ var googleOauthConfig = &oauth2.Config{
 	Endpoint:     google.Endpoint,
 }
 
-func validateToken(accessToken string) (bool, error) {
-	// Construct the tokeninfo URL with the access token
-	tokeninfoURL := fmt.Sprintf("https://www.googleapis.com/oauth2/v3/tokeninfo?access_token=%s", accessToken)
-
-	// Make a GET request to the tokeninfo endpoint
-	resp, err := http.Get(tokeninfoURL)
-	if err != nil {
-		return false, err
-	}
-	defer resp.Body.Close()
-
-	// Read the response body
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return false, err
-	}
-
-	// Parse the JSON response
-	var tokenInfo map[string]interface{}
-	err = json.Unmarshal(body, &tokenInfo)
-	if err != nil {
-		return false, err
-	}
-
-	// Check if the token is valid
-	if resp.StatusCode == http.StatusOK {
-		return true, nil
-	} else {
-		errorDescription := tokenInfo["error_description"].(string)
-		return false, fmt.Errorf("token validation failed: %s", errorDescription)
-	}
-}
-
-type BuzzForm struct {
-	A string `json:"a"`
-	P string `json:"p"`
-}
-
-func submitEndpoint(w http.ResponseWriter, r *http.Request) {
+func submitHandler(w http.ResponseWriter, r *http.Request) {
 
 	// Read the request body
 	var data BuzzForm
@@ -195,107 +116,67 @@ func submitEndpoint(w http.ResponseWriter, r *http.Request) {
 		log.Fatal(err)
 	}
 
-	// w.Header().Set("Content-Type", "text/html")
 	fmt.Fprint(w, html)
 
 }
 
-func oauthGoogleLogin(w http.ResponseWriter, r *http.Request) {
+func loginHandler(w http.ResponseWriter, r *http.Request) {
 	oauthState := generateStateOauthCookie(w)
 	host := getEnvOrDefault("REDIR", "http://localhost:8000")
-	redirectUrl := host + "/auth/google/callback"
+	redirectUrl := host + "/app"
 	googleOauthConfig.RedirectURL = redirectUrl
+
+	log.Printf("redirectUrl: %s", redirectUrl)
+
 	u := googleOauthConfig.AuthCodeURL(oauthState)
+
 	http.Redirect(w, r, u, http.StatusTemporaryRedirect)
 }
 
-func oauthGoogleCallback(w http.ResponseWriter, r *http.Request) {
-	// Read oauthState from Cookie
-	oauthState, _ := r.Cookie("oauthstate")
+func appHandler(w http.ResponseWriter, r *http.Request) {
 
-	if r.FormValue("state") != oauthState.Value {
-		log.Println("invalid oauth google state")
-		http.Redirect(w, r, "/", http.StatusTemporaryRedirect)
-		return
-	}
 	data, err := getUserDataFromGoogle(r.FormValue("code"))
 	if err != nil {
 		log.Println(err.Error())
 		http.Redirect(w, r, "/", http.StatusTemporaryRedirect)
 		return
 	}
-	dataBytes, err := json.Marshal(data)
+
+	// Step 1: Read the HTML template
+	htmlContent, err := os.ReadFile("./templates/app.html")
 	if err != nil {
-		log.Println(err.Error())
-		http.Redirect(w, r, "/", http.StatusTemporaryRedirect)
+		fmt.Println("Error reading file:", err)
 		return
 	}
-	// call AI -- TODO
-	http.FileServer(http.Dir("templates/"))
-	w.Write(dataBytes)
 
-}
-
-func generateStateOauthCookie(w http.ResponseWriter) string {
-	var expiration = time.Now().Add(20 * time.Minute)
-
-	b := make([]byte, 16)
-	rand.Read(b)
-	state := base64.URLEncoding.EncodeToString(b)
-	cookie := http.Cookie{Name: "oauthstate", Value: state, Expires: expiration}
-	http.SetCookie(w, &cookie)
-
-	return state
-}
-
-func getUserDataFromGoogle(code string) (Profile, error) {
-	// Use code to get token and get user info from Google.
-	host := getEnvOrDefault("REDIR", "http://localhost:8000")
-	redirectUrl := host + "/auth/google/callback"
-	googleOauthConfig.RedirectURL = redirectUrl
-	token, err := googleOauthConfig.Exchange(context.Background(), code)
+	// Step 2: Parse the HTML
+	doc, err := html.Parse(strings.NewReader(string(htmlContent)))
 	if err != nil {
-		return Profile{}, fmt.Errorf("code exchange wrong: %s", err.Error())
+		fmt.Println("Error parsing HTML:", err)
+		return
 	}
 
-	// Extract the OAuth2 token
-	accessToken := token.AccessToken
-	// refreshToken := token.RefreshToken/
-	// expiry := token.Expiry.RFC3339
+	// Step 3: Update the content of the div with a specific ID
+	targetID := "profile"
+	newContent := data.Name
+	updateDivContent(doc, targetID, newContent)
 
-	// log.Printf("Access Token: %s\n", accessToken)
-	// log.Printf( "Refresh Token: %s\n", refreshToken)
-	// log.Printf("Expiry: %s\n", expiry)
+	targetID = "a"
+	newContent = data.AccessToken
+	updateFormContent(doc, targetID, newContent)
 
-	response, err := http.Get(OAUTH_API + token.AccessToken)
+	// Step 4: Convert the updated HTML structure back to a string
+	var sb strings.Builder
+	err = html.Render(&sb, doc)
 	if err != nil {
-		return Profile{}, fmt.Errorf("failed getting user info: %s", err.Error())
+		fmt.Println("Error rendering HTML:", err)
+		return
 	}
-	defer response.Body.Close()
-	contents, err := io.ReadAll(response.Body)
-	if err != nil {
-		return Profile{}, fmt.Errorf("failed read response: %s", err.Error())
-	}
-	myProfile := Profile{}
-	err = json.Unmarshal(contents, &myProfile)
-	if err != nil {
-		return Profile{}, fmt.Errorf("invalid response: %s", err.Error())
-	}
-	myProfile.AccessToken = accessToken
-	return myProfile, nil
-	// return contents
-	// return append(contents, []byte(", { accessToken:"+accessToken+"}")...), nil
-}
 
-type Profile struct {
-	ID            string `json:"id"`
-	Email         string `json:"email"`
-	VerifiedEmail bool   `json:"verified_email"`
-	Name          string `json:"name"`
-	GivenName     string `json:"given_name"`
-	FamilyName    string `json:"family_name"`
-	Picture       string `json:"picture"`
-	Locale        string `json:"locale"`
-	Hd            string `json:"hd"`
-	AccessToken   string `json:"accessToken"`
+	// Set the HTTP response headers
+	w.Header().Set("Content-Type", "text/html; charset=utf-8")
+
+	// Write the updated HTML content as the HTTP response
+	fmt.Fprint(w, sb.String())
+
 }
