+++ app/templates/app.html
@@ -1,0 +1,114 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+
+    <title>Welcome to Buzz!</title>
+    <style>
+        .gsi-material-button {
+            -moz-user-select: none;
+            -webkit-user-select: none;
+            -ms-user-select: none;
+            -webkit-appearance: none;
+            background-color: #131314;
+            background-image: none;
+            border: 1px solid #747775;
+            -webkit-border-radius: 20px;
+            border-radius: 20px;
+            -webkit-box-sizing: border-box;
+            box-sizing: border-box;
+            color: #e3e3e3;
+            cursor: pointer;
+            font-family: 'Roboto', arial, sans-serif;
+            font-size: 14px;
+            height: 40px;
+            letter-spacing: 0.25px;
+            outline: none;
+            overflow: hidden;
+            padding: 0 12px;
+            position: relative;
+            text-align: center;
+            -webkit-transition: background-color .218s, border-color .218s, box-shadow .218s;
+            transition: background-color .218s, border-color .218s, box-shadow .218s;
+            vertical-align: middle;
+            white-space: nowrap;
+            width: auto;
+            max-width: 400px;
+            min-width: min-content;
+            border-color: #8e918f;
+        }
+
+        body {
+            background-color: black;
+            /* sets background color to black */
+            color: white;
+            text-align: center;
+            /* sets font color to white */
+        }
+
+        .container {
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            height: 100vh;
+            /* Adjust as needed */
+        }
+
+        .centered {
+            width: 50%;
+            /* Any fixed width */
+            margin-left: auto;
+            margin-right: auto;
+            border: 1px solid #000;
+            text-align: center;
+        }
+    </style>
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
+
+</head>
+
+<body class="centered">
+    <div id="profile"></div>
+    <img src="buzz.jpg" alt="Buzz App" width="20%">
+
+    <form id="myForm">
+        <label for="a">access token:</label><br>
+        <input type="text" id="a" name="a"><br>
+        <label for="p">phrase:</label><br>
+        <input type="text" id="p" name="p"><br><br>
+        <input type="button" class="gsi-material-button" value="🐝🤖✨" onclick="updateContent()">
+    </form>
+
+    <div id="updatedContent">
+    </div>
+
+    <script>
+        function updateContent() {
+            // Get form data
+            var formData = {
+                a: $("#a").val(),
+                p: $("#p").val()
+            };
+
+            // Make a POST request to the server
+            $.ajax({
+                type: "POST",
+                url: "/submit",
+                data: JSON.stringify(formData),
+                contentType: "application/json",
+                success: function (data) {
+                    // Update the content with the response
+                    $("#updatedContent").html(data);
+                },
+                error: function (jqXHR, textStatus, errorThrown) {
+                    console.error("Error:", textStatus, errorThrown);
+                }
+            });
+        }
+    </script>
+
+</body>
+
+</html>
